Added token auth so no hackable endpoints

+++ pages/api/deleteGoal.js
@@ -7,7 +7,28 @@ export default async function handler(req, res) {
   }
 
   try {
+    const authHeader = req.headers.authorization;
+    const token = authHeader?.split("Bearer ")[1].trim();
+    if (!token) {
+      throw new Error('no token ')
+    }
+
+    const payload = await accessTokenVerifier.verify(token);
+
     const data = req.body;
+    
+    const goal = await prisma.goal.findFirst({
+      where: {
+       id: data.id,
+       schedule: {
+        userId: payload.sub
+       }
+     }
+    });
+
+    if (!goal) {
+      return res.status(403).json({ error: 'Unauthorized' });
+    }
 
     await prisma.loggingOfMetric.deleteMany({
       where: {
@@ -17,7 +38,7 @@ export default async function handler(req, res) {
 
     await prisma.goal.delete({
       where: {
-        id: data.id
+        id: data.id,
       },
       include: {
         timeboxes: {
