+++ pages/api/deleteSchedule.js
@@ -1,5 +1,6 @@
 import prisma from "@/modules/prismaClient";
 import * as Sentry from "@sentry/nextjs";
+import { accessTokenVerifier } from "@/modules/cognitoVerifier";
 
 export default async function handler(req, res) {
   if (req.method !== 'POST') {
@@ -7,12 +8,20 @@ export default async function handler(req, res) {
   }
 
   try {
-    const data = req.body;
+    const authHeader = req.headers.authorization;
+    const token = authHeader?.split("Bearer ")[1].trim();
+    if (!token) {
+      throw new Error('no token ')
+    }
+
+    const payload = await accessTokenVerifier.verify(token);
+    
+    let data = req.body;
 
     await prisma.schedule.delete({
       where: {
         id: data.id,
-        userUUID: data.userUUID
+	userUUID: payload.sub
       },
       include: {
         timeboxes: true,
